requirements.txt
asgiref==3.3.1
Django==3.1.3
django-birthday==0.1.3
django-rest-multiple-models==2.1.3
djangorestframework==3.12.2
drf-writable-nested==0.6.2
generics==3.0.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
